[Refactor] refactoring create post provider by generator riverpod
